requirements.txt
numpy>=1.24,<2
pandas>=2.0,<2.2
opencv-python>=4.8
Pillow>=10.0
albumentations>=1.4
tqdm>=4.66
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
python-Levenshtein>=0.20
pytesseract>=0.3.10
easyocr>=1.7
torch>=2.0,<2.2
torchvision>=0.15,<0.17
jupyterlab>=4.0

